requirements.txt
flask == 1.0.2
flask_cors == 3.0.6
flask_caching == 1.4.0
flask_sqlalchemy == 2.3.2
flask_jwt_extended == 3.12.1
psycopg2 == 2.7.5
python-dotenv == 0.9.1
cerberus == 1.2
rauth == 0.7.3
redis == 2.10.6
celery == 4.2.1
celery-redbeat == 0.11.1
gunicorn == 19.9.0
